Tweaked authRequired to work with the new users table and created /me route that sends back the current user
